flask_app/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
fpdf2==2.7.6
Werkzeug==3.0.1
